# Conflicts:
#	audace-backend/src/main/java/com/equipe4/audace/AudaceApplication.java
#	audace-backend/src/main/java/com/equipe4/audace/model/application/Application.java
#	audace-backend/src/main/java/com/equipe4/audace/repository/ApplicationRepository.java
#	audace-backend/src/main/java/com/equipe4/audace/repository/offer/OfferRepository.java
#	audace-backend/src/main/java/com/equipe4/audace/service/EmployerService.java
#	audace-backend/src/main/java/com/equipe4/audace/service/ManagerService.java
#	audace-backend/src/main/java/com/equipe4/audace/service/StudentService.java
#	audace-backend/src/test/java/com/equipe4/audace/controller/EmployerControllerTest.java
#	audace-backend/src/test/java/com/equipe4/audace/controller/ManagerControllerTest.java
#	audace-backend/src/test/java/com/equipe4/audace/service/EmployerServiceTest.java
#	audace-backend/src/test/java/com/equipe4/audace/service/ManagerServiceTest.java
#	audace-backend/src/test/java/com/equipe4/audace/service/StudentServiceTest.java
#	audace-frontend/src/components/ApplicationsList/ApplicationRow/index.tsx
#	audace-frontend/src/components/OffersList/OfferRow/OfferButtons/StudentButtons/index.tsx
#	audace-frontend/src/components/OffersList/OfferRow/OfferModal/index.tsx
#	audace-frontend/src/services/cvService.ts
#	audace-frontend/src/services/offerService.ts
#	audace-frontend/src/services/studentApplicationService.ts
#	audace-frontend/src/views/Student/StudentView/index.tsx

+++ audace-backend/src/main/java/com/equipe4/audace/service/EmployerService.java
@@ -1,43 +1,45 @@
 package com.equipe4.audace.service;
 
-import com.equipe4.audace.dto.application.ApplicationDTO;
 import com.equipe4.audace.dto.EmployerDTO;
-import com.equipe4.audace.dto.department.DepartmentDTO;
+import com.equipe4.audace.dto.application.ApplicationDTO;
 import com.equipe4.audace.dto.offer.OfferDTO;
-import com.equipe4.audace.model.application.Application;
 import com.equipe4.audace.model.Employer;
-import com.equipe4.audace.model.department.Department;
+import com.equipe4.audace.model.application.Application;
 import com.equipe4.audace.model.offer.Offer;
-import com.equipe4.audace.repository.ApplicationRepository;
+import com.equipe4.audace.model.session.OfferSession;
+import com.equipe4.audace.model.session.Session;
 import com.equipe4.audace.repository.EmployerRepository;
+import com.equipe4.audace.repository.application.ApplicationRepository;
 import com.equipe4.audace.repository.department.DepartmentRepository;
 import com.equipe4.audace.repository.offer.OfferRepository;
 import com.equipe4.audace.repository.security.SaltRepository;
+import com.equipe4.audace.repository.session.OfferSessionRepository;
+import com.equipe4.audace.utils.SessionManipulator;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
+import java.util.NoSuchElementException;
+
 
 @Service
 public class EmployerService extends GenericUserService<Employer> {
     private final EmployerRepository employerRepository;
     private final OfferRepository offerRepository;
     private final ApplicationRepository applicationRepository;
-
+    private final OfferSessionRepository offerSessionRepository;
+    private final SessionManipulator sessionManipulator;
     private final DepartmentRepository departmentRepository;
 
-    public EmployerService(
-            SaltRepository saltRepository,
-            EmployerRepository employerRepository,
-            OfferRepository offerRepository,
-            ApplicationRepository applicationRepository,
-            DepartmentRepository departmentRepository
-    ) {
+    public EmployerService(SaltRepository saltRepository, EmployerRepository employerRepository, OfferRepository offerRepository,
+                           OfferSessionRepository offerSessionRepository, SessionManipulator sessionManipulator,
+                           ApplicationRepository applicationRepository, DepartmentRepository departmentRepository) {
         super(saltRepository);
         this.employerRepository = employerRepository;
         this.offerRepository = offerRepository;
+        this.offerSessionRepository = offerSessionRepository;
+        this.sessionManipulator = sessionManipulator;
         this.applicationRepository = applicationRepository;
         this.departmentRepository = departmentRepository;
     }
@@ -47,7 +49,6 @@ public class EmployerService extends GenericUserService<Employer> {
         if (employerDTO == null) throw new IllegalArgumentException("Employer cannot be null");
 
         Optional<Employer> employerOptional = employerRepository.findByEmail(employerDTO.getEmail());
-
         if(employerOptional.isPresent()) throw new IllegalArgumentException("Email already in use");
 
         Employer newEmployer = employerDTO.fromDTO();
@@ -63,51 +64,49 @@ public class EmployerService extends GenericUserService<Employer> {
         return employerRepository.findAll().stream().map(Employer::toDTO).toList();
     }
 
+    @Transactional
     public Optional<OfferDTO> createOffer(OfferDTO offerDTO){
         if(offerDTO == null) throw new IllegalArgumentException("Offer cannot be null");
 
-        Employer employer = employerRepository.findById(offerDTO.getEmployerId()).orElseThrow();
-        Department department = departmentRepository.findByCode(offerDTO.getDepartmentCode()).orElseThrow();
+        Session session = sessionManipulator.getCurrentSession();
 
-        Offer offer = offerDTO.fromDTO();
-        offer.setEmployer(employer);
-        offer.setDepartment(department);
+        offerDTO.setOfferStatus(Offer.OfferStatus.PENDING);
 
+        Offer offer = offerRepository.save(offerDTO.fromDTO());
+
+        offerSessionRepository.save(new OfferSession(null, offer, session));
+
+        return Optional.of(offer.toDTO());
+    }
+
+    @Transactional
+    public List<OfferDTO> findAllOffersByEmployerIdAndSessionId(Long employerId, Long sessionId){
+        Employer employer = employerRepository.findById(employerId).orElseThrow(() -> new NoSuchElementException("Employer not found"));
+        List<Offer> offers = offerRepository.findAllByEmployer(employer);
+
+        return sessionManipulator.removeOffersNotInSession(offers, sessionId).stream().map(Offer::toDTO).toList();
+    }
+
+    @Transactional
+    public Optional<OfferDTO> updateOffer(OfferDTO offerDTO){
+        Offer offer = offerRepository.findById(offerDTO.getId()).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        if (!sessionManipulator.isOfferInCurrentSession(offer)) throw new IllegalStateException("Offer is not in current session");
+        offer = offerDTO.fromDTO();
         return Optional.of(offerRepository.save(offer).toDTO());
     }
 
-    public Optional<OfferDTO> findOfferById(Long offerId){
-        return offerRepository.findById(offerId).map(Offer::toDTO);
-    }
-
-    public List<DepartmentDTO> findAllDepartments(){
-        return departmentRepository.findAll().stream().map(Department::toDTO).toList();
-    }
-
-    public List<OfferDTO> findAllOffersByEmployerId(Long employerId){
-        Employer employer = employerRepository.findById(employerId).orElseThrow();
-        return offerRepository.findAllByEmployer(employer).stream().map(Offer::toDTO).toList();
-    }
-    public Optional<OfferDTO> updateOffer(OfferDTO offerDTO) {
-        Offer offer = offerRepository.findById(offerDTO.getId()).orElseThrow(() -> new NoSuchElementException("Offer with ID " + offerDTO.getId() + " not found."));
-
-        offer.setTitle(offerDTO.getTitle());
-        offer.setDescription(offerDTO.getDescription());
-        offer.setInternshipStartDate(offerDTO.getInternshipStartDate());
-        offer.setInternshipEndDate(offerDTO.getInternshipEndDate());
-        offer.setOfferEndDate(offerDTO.getOfferEndDate());
-        offer.setAvailablePlaces(offerDTO.getAvailablePlaces());
-        offer.setDepartment(departmentRepository.findByCode(offerDTO.getDepartmentCode()).orElseThrow());
-
-        return Optional.of(offerRepository.save(offer).toDTO());
-    }
-
-
-
+    @Transactional
     public void deleteOffer(Long offerId){
-        Offer offer = offerRepository.findById(offerId).orElseThrow();
+        Offer offer = offerRepository.findById(offerId).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        if (!sessionManipulator.isOfferInCurrentSession(offer)) throw new IllegalStateException("Offer is not in current session");
         offerRepository.delete(offer);
     }
+
+    public List<ApplicationDTO> findAllApplicationsByEmployerIdAndOfferId(Long employerId, Long offerId) {
+        Offer offer = offerRepository.findByEmployerIdAndId(employerId, offerId).orElseThrow(() -> new NoSuchElementException("Offer not found"));
+        return applicationRepository.findAllByOffer(offer).stream().map(Application::toDTO).toList();
+    }
+
     @Transactional
     public Optional<ApplicationDTO> acceptApplication(Long employerId, Long applicationId) {
         return setApplicationStatus(employerId, applicationId, Application.ApplicationStatus.ACCEPTED);
@@ -118,20 +117,18 @@ public class EmployerService extends GenericUserService<Employer> {
     }
 
     private Optional<ApplicationDTO> setApplicationStatus(Long employerId, Long applicationId, Application.ApplicationStatus applicationStatus) {
-        Application application = applicationRepository.findById(applicationId).orElseThrow();
-        if (!application.getOffer().getEmployer().getId().equals(employerId)) {
-            throw new IllegalArgumentException("Employer does not own this application");
-        }
+        Application application = applicationRepository.findById(applicationId).orElseThrow(() -> new NoSuchElementException("Application not found"));
+
+        if (!application.getOffer().getEmployer().getId().equals(employerId)) throw new IllegalArgumentException("Employer does not own this application");
+
         if (applicationStatus == Application.ApplicationStatus.ACCEPTED) {
             Offer offer = application.getOffer();
-            Long placesUsed = applicationRepository
-                    .countApplicationsByApplicationStatusAndOffer(
-                            Application.ApplicationStatus.ACCEPTED,
-                            offer
-                    );
-            if (placesUsed >= offer.getAvailablePlaces()) {
-                throw new IllegalArgumentException("No more places available");
-            }
+            int availablePlaces = offer.getAvailablePlaces();
+            if (availablePlaces <= 0) throw new IllegalArgumentException("No more places available");
+
+            offer.setAvailablePlaces(availablePlaces - 1);
+            offerRepository.save(offer);
+
         }
         application.setApplicationStatus(applicationStatus);
         applicationRepository.save(application);
